handling json

+++ src/main/java/org/app/auth/controllers/ImageController.java
@@ -1,5 +1,7 @@
 package org.app.auth.controllers;
 
+import com.google.gson.Gson;
+import org.app.auth.POJOs.Prediction;
 import org.app.auth.entities.Image;
 import org.app.auth.services.ImageService;
 import org.springframework.context.annotation.Lazy;
@@ -18,8 +20,8 @@ public class ImageController {
     @PostMapping("/auth/saveimg")
     @ResponseBody
     public String createImage(@RequestParam("filename") String filename,
-                              @RequestParam("score") String score,
-                              @RequestParam("acc") String acc,
+                              @RequestParam("score") Double score,
+                              @RequestParam("acc") Double acc,
                               @RequestParam("uname") String username,
                               @RequestParam("link") String link) {
         try {
@@ -39,11 +41,14 @@ public class ImageController {
     @PostMapping("/auth/imgpred")
     @ResponseBody
     public String getImgPrediction(@RequestParam("filename") String filename,
-                                   @RequestParam("username") String username) {
+                                       @RequestParam("username") String username) {
         try {
-            return imageService.getPrediction(filename, username);
+            Prediction prediction =
+                    imageService.getPrediction(filename, username);
+            String json = new Gson().toJson(prediction);
+            return json;
         } catch (Exception e) {
-            return "failed";
+            return "fail";
         }
     }
 
@@ -51,12 +56,14 @@ public class ImageController {
     @ResponseBody
     public void delImage(@RequestParam("filename") String filename,
                          @RequestParam("username") String username) {
-        String APP_PATH = "/home/bananapp/auth";
+        String systemuser = System.getProperty("user.name");
+        String APP_PATH = "/home/"+systemuser+"/bananapp/auth";
 
         imageService.delImage(filename, username);
 
         String filepath = APP_PATH + File.separator + username + File.separator + filename;
         File file = new File(filepath);
-        file.delete();
+        System.out.println(file.exists());
+        if (file.exists()) file.delete();
     }
 }
